+++ src/components/card/CardNumberInput.tsx
@@ -1,8 +1,7 @@
 import { Fragment } from "react/jsx-runtime";
-import Input from "./Input";
-import ValidationMessage from "./ValidationMessage";
-import { useBlur, useFocusNext } from "../hooks";
-import { cardValidator } from "../domain";
+import { useBlur, useFocusNext } from "../../hooks";
+import { cardValidator } from "../../domain";
+import { Input, ValidationMessage } from "../common";
 
 interface CardNumberInputProps {
   cardNumber: string[];
