+++ src/components/common/index.ts
@@ -1,0 +1,7 @@
+export { default as Button } from "./Button";
+export { default as Title } from "./Title";
+export { default as Input } from "./Input";
+export { default as Modal } from "./Modal";
+export { default as Icon } from "./Icon";
+export { default as Tooltip } from "./Tooltip";
+export { default as ValidationMessage } from "./ValidationMessage";
