+++ src/components/index.ts
@@ -1,8 +1,2 @@
-export { default as Card } from "./Card";
-export { default as Modal } from "./Modal";
-export { default as Icon } from "./Icon";
 export { default as CardInfoEdit } from "./CardInfoEdit";
 export { default as CardNicknameEdit } from "./CardNicknameEdit";
-export { default as Button } from "./Button";
-export { default as Title } from "./Title";
-export { default as Input } from "./Input";
